Refactor index component UI styling and background color
Simplify the index component HTML by removing inline style for the
'stats' image. Enhance the index component styling by updating the
'stats' ID selector in the SCSS file to center the image horizontally
and limit its width to 90% of the container. Also, change the background
color of the application to a darker shade of gray. These changes
improve the visual appearance and maintainability of the index
component.