[gitignore] Removed /dist directory to prevent tracking compiled output
Previously, the /dist directory was being tracked, which led to
unnecessary bloating and potential confusion. By removing it from the
.gitignore file, we ensure that the compiled output is not included in
version control. This optimizes the repository size and avoids
cluttering the commit history with compiled artifacts. It also aligns
with best practices for ignoring files with Git, as recommended by
GitHub.

Issue #123